requirements.txt
ipython
matplotlib
mdai
numpy
opencv-python
opencv-python-headless
pandas
pillow
python-dotenv
scikit-image
scipy
torch
torchvision
tqdm